requirements.txt
Django==4.1.3
psycopg2-binary==2.9.6
flake8==6.0.0
black==23.3
isort==5.12.0
gunicorn==20.1.0
uvicorn==0.21.1
whitenoise==6.2.0
django-filter==23.2
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
Pillow==9.5.0
django-storages==1.13.2
boto3==1.28.40
cryptography==41.0.3
# django-dbbackup==4.0.2
# demucs==4.0.1
# pydub==0.25.1
django-cors-headers==4.2.0
djangorestframework-api-key==2.3.0
python-magic==0.4.27
drf-extra-fields==3.7.0
requests==2.31.0
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
imutils==0.5.4
ffmpeg-python==0.2.0